+++ playground/vuxe.config.ts
@@ -1,0 +1,4 @@
+import { defineConfig } from 'vuxe'
+
+export default defineConfig({
+})